Apply inversion of dependencies on Outputable
Also include a mechanism (python 3.6) for registering all the
classes inheriting from Outputable in the mother class. This could
enable removing all the children classes from the Runner in the
long term. For this commit, it is not complete since the order of
the children classes could be messed up.