Add a wizard for helping the renaming and merging of author names
This solution does not work currently, because it writes data into
the .mailmap file of a repository (sthg not great in itself), and
these data are not honored with the next call of gitrepository.
Then, there needs to be added a mechanism within gitrepository to
handle the renaming and merging of the names.