Create a second TestOutput class mimicking BlameOutput
All files and code are currently in temporary locations. This is an
intermediary commit, and it remains to check if the code can be
refactored more properly. Remarks :

- the JS for both outputs are now isolated, no variable from the first
can go to the second.

- the colors are problematic, because they may be different between the
outputs. A solution would be to make the python compute the colors, and
then send them to the outputs.

- some tests are not passing for this commit