Assign a unique color to each author
The color is assigned in the Python code, instead of the HTML, which
could be seen as not modular. In theory, it should be possible to make
the HTML code contain a list of the authors, and compute a map of colors.
I decided that the color information could be useful in other sorts of
outputs.

As a matter of fact, this commit fixes a bug where the colors assigned
to the authors could differ in the outputs.