Allow the specification of extensions with regexps
This commit allows to pass options such as -f ".*\.c".
This looks inferior to the previous specification, but not
it becomes possible to consider files without extensions in
the repository, such as "Makefile".

Some comments on the code : the extensions module is, IMHO, flawed.
It contains two lists of extensions, one of them being allowed, and
one of them being encountered during the analysis. This commit copies
and considers only the filter in filtering for the allowed ones.

Maybe, in certain cases, there is something to see in the extensions
that have been encountered and not specified by configuration. Maybe.
But first, extensions are a flawed way of distnguishing files, regexps
are natural and more expressive (thin of "Mazfile"). Second, if the
interesting files are specified by config, well, the extensions
encountered during analysis are not enthralling. Last, in the case where
the extensions are not specified, I believe that all files should be
marked as interesting (and not only those allowed in the program).

I understand that some parts of the analysis are done depending on the
extensions of the files. Yeah. But basically, most file types are not
handled, so this is not a serious argument. I believe that this
mechanism should be removed from the code and simplified.