Reinstate the filters for excluded file names
This commit is the last of series to reinstate the intial behavior for accepted
and excluded file names. Hopefully, it should be much more flexible. Now, we
still have a list of extensions that are accepted by default. This list can be
modified on the command-line with the "-f" option and the "-x" and a pattern
"file_in:<regexp>". These are called "positive" patterns. Then, one can also
add other "-x" patterns such as "file_out<regexp>", e.g "negative" patterns.

The files that are considered during an inspection are those that belong to at
least one positive pattern, and to no negative pattern.

For reference, the other filters are just "negative" filters. But it should be
easy to implement a large set of filters if someone required it. One that I see
could be interesting would be files of a certain length.