Initial support for a "-b" option specifying the branch to inspect
The idea comes originally from another fork of gitinspector by 'mawer111',
and could effectively be practical for analyzing specific developments
done inside branches. Well, I have scanned the different calls to git
within the code, and hopefully, this should cover them all. For the moment,
I am quite content with the fact that this does not directly induce bugs.
This shall be tests with more specific repositories, that I don't have with
me at the moment.