Add a "--legacy" option for old outputs that have been superseded
Maybe it's a bit harsh, but these outputs are not maintained and
duplicate the information in the results. For the moment, I only
removed the TimelineOutput, because the ActivityOutput gives
IMHO a better vision of the responsibilities. It is still possible
to activate with the "--legacy" option.

For the moment, the ResponsibilitiesOutput is still default, but
the OwnershipsOutput gives a good representation of the same data.