Create a menu system for the HTML output
This involves a modification of the totality of the templates, the js
and some css. Unfortunately, I noticed afterwards that I had not ported
*all* the outputs to the new system, so this commit also contains these
transformations. There is a huge need of cleaning the js and the css,
but for the moment we need testing this new menu system on different
browsers.