Disable "blames" computation for all branches
This is problematic. As much as it is possible to consider the tip of
a branch, all files inside and the blames that are associated, when
setting "--all" to the branches, this becomes irrelevant. In fact,
there can be different blames for the same file in different branches.

This is a bit upsetting, but at least, this corrects a bug where the
blames were computed only for the HEAD of the current branch of the
repository, and not the branch set in the options.