Apply an heuristic to compute blames on all branches
Originally, this reflexion (this particular branch) came from the fact
that it is difficult to analyze all the commits in a repository on just
one branch. The "-b" setting made it easier to select the branch, but
the blames were originally computed on the current branch of the repository,
which made them unaccurate.

On the other hand, it was decided that the most sensible default for that
branch would be the "--all" flag, but this made the blames all the more
difficult to compute.

This commit introduces a heuristic for computing blames on all branches :
if a file appears to have been modified on multiple branches, then the
blames that count are the ones for the most recent file.

This heuristic is not 100%-accurate. Yet it looks like a sensible way to
do things. When setting a concrete branch with the -b option, the blames
are computed as accurately as before. Maybe some people would prefer that
the default for -b were "master" instead of "--all". For the study of
students repositories, it was decided to prefer "--all".