Make the commits honot the "--all" branch argument
For the moment, the commits are considered in all the branches, but
they are summed. I have experienced with sets, but this fails, because
commits are handled as consecutive pairs. I am not entirely certain
that it counts the commits correctly, since all branches have a fair
chance of having the same base, and therefore counting these diffs
multiple times. Wee, it will suffice for the moment.