Remove the "--no-merges" when computing commit chunks
This line has no effet in practice, since when handling one chunk that
corresponds to a merge, the chunk contains no acceptable file, and is
therefore not considered by gitinspector. Yes, they could be removed
completely, but if we ever consider counting the works of the rejected
commits somewhere, this better stay in place.